Add github-search-pronto.user.js
User script that triggers a search within a repository immediately upon
pressing enter.

This used to be the normal functionality, but ever since GitHub added
the search drop-down to enable a choice of search scopes, they don't let
you press enter to search until the drop-down list has populated. That
typically takes a second or two, which means you have to press enter
again after a second in order to actually search for something. So
stupid.

Here we search right away unless the search drop-down is visible and has
been populated.